[sdk] Unify Limit type to int across all service options
GetLogsOptions.Limit was int64 while ListBackupsOptions.Limit and
ListRestoresOptions.Limit were int. Standardize to int since all
practical counts are small. Cast to int64 at the PBM call boundary.

+++ sdk/log.go
@@ -66,7 +66,7 @@ type LogService interface {
 // GetLogsOptions controls filtering and pagination for log retrieval.
 type GetLogsOptions struct {
 	// Limit is the maximum number of log entries to return. Zero means no limit.
-	Limit int64
+	Limit int
 
 	// Severity is the minimum severity level to include. PBM's severity filter
 	// includes all entries at the given level and above (e.g. Info includes
